修改数据库字段 增加孕周文字格式30周+4

+++ library/src/main/java/cn/ihealthbaby/weitaixin/library/data/database/dao/RecordBusinessDao.java
@@ -191,7 +191,7 @@ public class RecordBusinessDao {
 		Record query = query(record);
 		Integer uploadState = record.getUploadState();
 		Date recordStartTime = record.getRecordStartTime();
-		Long duration = record.getDuration();
+		Integer duration = record.getDuration();
 		String recordData = record.getRecordData();
 		String soundPath = record.getSoundPath();
 		Integer feelingId = record.getFeelingId();
